+++ app/models/pay/charge.rb
@@ -23,20 +23,8 @@ module Pay
     store_accessor :data, :username # Venmo
     store_accessor :data, :bank
 
-    store_accessor :data, :amount_captured
-    store_accessor :data, :invoice_id
-    store_accessor :data, :payment_intent_id
-    store_accessor :data, :period_start
-    store_accessor :data, :period_end
-    store_accessor :data, :line_items
-    store_accessor :data, :subtotal # subtotal amount in cents
-    store_accessor :data, :tax # total tax amount in cents
-    store_accessor :data, :discounts # array of discount IDs applied to the Stripe Invoice
-    store_accessor :data, :total_discount_amounts # array of discount details
-    store_accessor :data, :total_tax_amounts # array of tax details for each jurisdiction
-    store_accessor :data, :credit_notes # array of credit notes for the Stripe Invoice
-    store_accessor :data, :refunds # array of refunds
-    store_accessor :data, :balance_transaction
+    store_accessor :data, :subtotal
+    store_accessor :data, :tax
 
     # Helpers for payment processors
     %w[braintree stripe paddle_billing paddle_classic lemon_squeezy fake_processor].each do |processor_name|
@@ -54,10 +42,6 @@ module Pay
     def sync!(**options)
       self.class.sync(processor_id, **options)
       reload
-    end
-
-    def captured?
-      amount_captured > 0
     end
 
     def refunded?
@@ -117,10 +101,6 @@ module Pay
       else
         payment_method_type&.titleize
       end
-    end
-
-    def line_items
-      Array.wrap(super)
     end
   end
 end