+++ app/models/pay/stripe/charge.rb
@@ -1,11 +1,21 @@
 module Pay
   module Stripe
     class Charge < Pay::Charge
+      EXPAND = ["balance_transaction", "refunds"]
+
+      delegate :amount_captured, :payment_intent, to: :stripe_object
+
+      store_accessor :data, :stripe_invoice
       store_accessor :data, :stripe_receipt_url
+
+      def self.sync_payment_intent(id, stripe_account: nil)
+        payment_intent = ::Stripe::PaymentIntent.retrieve({id: id}, {stripe_account: stripe_account}.compact)
+        sync(payment_intent.latest_charge, stripe_account: stripe_account)
+      end
 
       def self.sync(charge_id, object: nil, stripe_account: nil, try: 0, retries: 1)
         # Skip loading the latest charge details from the API if we already have it
-        object ||= ::Stripe::Charge.retrieve({id: charge_id, expand: ["balance_transaction", "invoice.total_discount_amounts.discount", "invoice.total_tax_amounts.tax_rate", "refunds"]}, {stripe_account: stripe_account}.compact)
+        object ||= ::Stripe::Charge.retrieve({id: charge_id, expand: EXPAND}, {stripe_account: stripe_account}.compact)
         if object.customer.blank?
           Rails.logger.debug "Stripe Charge #{object.id} does not have a customer"
           return
@@ -17,68 +27,34 @@ module Pay
           return
         end
 
-        refunds = []
-        object.refunds.auto_paging_each { |refund| refunds << refund }
-
         payment_method = object.payment_method_details.try(object.payment_method_details.type)
         attrs = {
+          object: object.to_hash,
           amount: object.amount,
-          amount_captured: object.amount_captured,
           amount_refunded: object.amount_refunded,
           application_fee_amount: object.application_fee_amount,
-          balance_transaction: object.balance_transaction,
           bank: payment_method.try(:bank_name) || payment_method.try(:bank), # eps, fpx, ideal, p24, acss_debit, etc
           brand: payment_method.try(:brand)&.capitalize,
           created_at: Time.at(object.created),
           currency: object.currency,
-          discounts: [],
           exp_month: payment_method.try(:exp_month).to_s,
           exp_year: payment_method.try(:exp_year).to_s,
           last4: payment_method.try(:last4).to_s,
-          line_items: [],
           metadata: object.metadata,
-          payment_intent_id: object.payment_intent,
           payment_method_type: object.payment_method_details.type,
           stripe_account: pay_customer.stripe_account,
-          stripe_receipt_url: object.receipt_url,
-          total_tax_amounts: [],
-          refunds: refunds.sort_by! { |r| r["created"] }
+          stripe_receipt_url: object.receipt_url
         }
 
         # Associate charge with subscription if we can
-        if object.invoice
-          invoice = (object.invoice.is_a?(::Stripe::Invoice) ? object.invoice : ::Stripe::Invoice.retrieve({id: object.invoice, expand: ["total_discount_amounts.discount", "total_tax_amounts.tax_rate"]}, {stripe_account: stripe_account}.compact))
-          attrs[:invoice_id] = invoice.id
-          attrs[:subscription] = pay_customer.subscriptions.find_by(processor_id: invoice.subscription)
-
-          attrs[:period_start] = Time.at(invoice.period_start)
-          attrs[:period_end] = Time.at(invoice.period_end)
+        invoice_payments = ::Stripe::InvoicePayment.list({payment: {type: :payment_intent, payment_intent: object.payment_intent}, status: :paid, expand: ["data.invoice.total_discount_amounts.discount"]}, {stripe_account: stripe_account}.compact)
+        if invoice_payments.any? && (invoice = invoice_payments.first.invoice)
+          attrs[:stripe_invoice] = invoice.to_hash
           attrs[:subtotal] = invoice.subtotal
-          attrs[:tax] = invoice.tax
-          attrs[:discounts] = invoice.discounts
-          attrs[:total_tax_amounts] = invoice.total_tax_amounts.map(&:to_hash)
-          attrs[:total_discount_amounts] = invoice.total_discount_amounts.map(&:to_hash)
-
-          invoice.lines.auto_paging_each do |line_item|
-            # Currency is tied to the charge, so storing it would be duplication
-            attrs[:line_items] << {
-              id: line_item.id,
-              description: line_item.description,
-              price_id: line_item.price&.id,
-              quantity: line_item.quantity,
-              unit_amount: line_item.price&.unit_amount,
-              amount: line_item.amount,
-              discounts: line_item.discounts,
-              tax_amounts: line_item.tax_amounts,
-              proration: line_item.proration,
-              period_start: Time.at(line_item.period.start),
-              period_end: Time.at(line_item.period.end)
-            }
+          attrs[:tax] = invoice.total - invoice.total_excluding_tax.to_i
+          if (subscription = invoice.parent.try(:subscription_details).try(:subscription))
+            attrs[:subscription] = pay_customer.subscriptions.find_by(processor_id: subscription)
           end
-        # Charges without invoices
-        else
-          attrs[:period_start] = Time.at(object.created)
-          attrs[:period_end] = Time.at(object.created)
         end
 
         # Update or create the charge
@@ -90,16 +66,14 @@ module Pay
         end
       rescue ActiveRecord::RecordInvalid, ActiveRecord::RecordNotUnique
         try += 1
-        if try <= retries
-          sleep 0.1
-          retry
-        else
-          raise
-        end
+        raise unless try <= retries
+
+        sleep 0.1
+        retry
       end
 
       def api_record
-        ::Stripe::Charge.retrieve({id: processor_id, expand: ["customer", "invoice.subscription"]}, stripe_options)
+        ::Stripe::Charge.retrieve({id: processor_id, expand: EXPAND}, stripe_options)
       rescue ::Stripe::StripeError => e
         raise Pay::Stripe::Error, e
       end
@@ -116,7 +90,7 @@ module Pay
       def refund!(amount_to_refund, **options)
         amount_to_refund ||= amount
 
-        if invoice_id.present?
+        if stripe_invoice.present?
           description = options.delete(:description) || I18n.t("pay.refund")
           lines = [{type: :custom_line_item, description: description, quantity: 1, unit_amount: amount_to_refund}]
           credit_note!(**options.merge(refund_amount: amount_to_refund, lines: lines))
@@ -130,8 +104,9 @@ module Pay
 
       # Adds a credit note to a Stripe Invoice
       def credit_note!(**options)
-        raise Pay::Stripe::Error, "no Stripe invoice_id on Pay::Charge" if invoice_id.blank?
-        ::Stripe::CreditNote.create({invoice: invoice_id}.merge(options), stripe_options)
+        raise Pay::Stripe::Error, "no Stripe Invoice on Pay::Charge" if stripe_invoice.blank?
+
+        ::Stripe::CreditNote.create({invoice: stripe_invoice.id}.merge(options), stripe_options)
       rescue ::Stripe::StripeError => e
         raise Pay::Stripe::Error, e
       end
@@ -141,11 +116,26 @@ module Pay
       # capture
       # capture(amount_to_capture: 15_00)
       def capture(**options)
-        raise Pay::Stripe::Error, "no payment_intent_id on charge" unless payment_intent_id.present?
-        ::Stripe::PaymentIntent.capture(payment_intent_id, options, stripe_options)
+        raise Pay::Stripe::Error, "no payment_intent on charge" unless payment_intent.present?
+
+        ::Stripe::PaymentIntent.capture(payment_intent, options, stripe_options)
         self.class.sync(processor_id)
       rescue ::Stripe::StripeError => e
         raise Pay::Stripe::Error, e
+      end
+
+      def captured?
+        amount_captured > 0
+      end
+
+      def stripe_invoice
+        if (value = data.dig("stripe_invoice"))
+          ::Stripe::Invoice.construct_from(value)
+        end
+      end
+
+      def stripe_object
+        ::Stripe::Charge.construct_from(object) if object?
       end
 
       private