+++ app/models/pay/stripe/subscription.rb
@@ -23,6 +23,7 @@ module Pay
         end
 
         attributes = {
+          object: object.to_hash,
           application_fee_percent: object.application_fee_percent,
           created_at: Time.at(object.created),
           processor_plan: object.items.first.price.id,
@@ -30,12 +31,11 @@ module Pay
           status: object.status,
           stripe_account: pay_customer.stripe_account,
           metadata: object.metadata,
-          subscription_items: [],
           metered: false,
           pause_behavior: object.pause_collection&.behavior,
           pause_resumes_at: (object.pause_collection&.resumes_at ? Time.at(object.pause_collection&.resumes_at) : nil),
-          current_period_start: (object.current_period_start ? Time.at(object.current_period_start) : nil),
-          current_period_end: (object.current_period_end ? Time.at(object.current_period_end) : nil)
+          current_period_start: (object.items.first.current_period_start ? Time.at(object.items.first.current_period_start) : nil),
+          current_period_end: (object.items.first.current_period_end ? Time.at(object.items.first.current_period_end) : nil)
         }
 
         # Subscriptions that have ended should have their trial ended at the
@@ -49,13 +49,9 @@ module Pay
           attributes[:trial_ends_at] = Time.at(trial_ended_at)
         end
 
-        # Record subscription items to db
         object.items.auto_paging_each do |subscription_item|
-          if !attributes[:metered] && (subscription_item.to_hash.dig(:price, :recurring, :usage_type) == "metered")
-            attributes[:metered] = true
-          end
-
-          attributes[:subscription_items] << subscription_item.to_hash.slice(:id, :price, :metadata, :quantity)
+          next if attributes[:metered]
+          attributes[:metered] = true if subscription_item.price.try(:recurring).try(:usage_type) == "metered"
         end
 
         attributes[:ends_at] = if object.ended_at
@@ -66,7 +62,7 @@ module Pay
           Time.at(object.cancel_at)
         elsif object.cancel_at_period_end
           # Subscriptions cancelling in the future
-          Time.at(object.current_period_end)
+          Time.at(object.items.first.current_period_end)
         end
 
         # Sync payment method if directly attached to subscription
@@ -87,7 +83,7 @@ module Pay
           # Any other pause status (or no pause at all) should have nil for start
           if pay_subscription.pause_behavior != attributes[:pause_behavior]
             attributes[:pause_starts_at] = if attributes[:pause_behavior] == "void"
-              Time.at(object.current_period_end)
+              Time.at(object.items.first.current_period_end)
             end
           end
 
@@ -102,8 +98,11 @@ module Pay
         pay_subscription.api_record = object
 
         # Sync the latest charge if we already have it loaded (like during subscrbe), otherwise, let webhooks take care of creating it
-        if (charge = object.try(:latest_invoice).try(:charge)) && charge.try(:status) == "succeeded"
-          Pay::Stripe::Charge.sync(charge.id, stripe_account: pay_subscription.stripe_account)
+        if (invoice = object.try(:latest_invoice))
+          Array(invoice.try(:payments)).each do |invoice_payment|
+            next unless invoice_payment.status == "paid"
+            Pay::Stripe::Charge.sync_payment_intent(invoice_payment.payment.payment_intent, stripe_account: pay_subscription.stripe_account)
+          end
         end
 
         pay_subscription
@@ -121,14 +120,18 @@ module Pay
       def self.expand_options
         {
           expand: [
+            "discounts",
             "default_payment_method",
             "pending_setup_intent",
-            "latest_invoice.payment_intent",
-            "latest_invoice.charge",
-            "latest_invoice.total_discount_amounts.discount",
-            "latest_invoice.total_tax_amounts.tax_rate"
+            "latest_invoice.confirmation_secret",
+            "latest_invoice.payments",
+            "latest_invoice.total_discount_amounts.discount"
           ]
         }
+      end
+
+      def stripe_object
+        ::Stripe::Subscription.construct_from(object)
       end
 
       def api_record(**options)
@@ -137,7 +140,7 @@ module Pay
 
       # Returns a SetupIntent or PaymentIntent client secret for the subscription
       def client_secret
-        api_record&.pending_setup_intent&.client_secret || api_record&.latest_invoice&.payment_intent&.client_secret
+        api_record&.pending_setup_intent&.client_secret || api_record&.latest_invoice&.confirmation_secret&.client_secret
       end
 
       # Sets the default_payment_method on a subscription
@@ -160,7 +163,7 @@ module Pay
           cancel_now!
         else
           @api_record = ::Stripe::Subscription.update(processor_id, {cancel_at_period_end: true}.merge(expand_options), stripe_options)
-          update(ends_at: (on_trial? ? trial_ends_at : Time.at(@api_record.current_period_end)))
+          update(ends_at: (on_trial? ? trial_ends_at : Time.at(@api_record.items.first.current_period_end)))
         end
       rescue ::Stripe::StripeError => e
         raise Pay::Stripe::Error, e
@@ -184,7 +187,7 @@ module Pay
       # For a subscription with a single item, we can update the subscription directly if no SubscriptionItem ID is available
       # Otherwise a SubscriptionItem ID is required so Stripe knows which entry to update
       def change_quantity(quantity, **options)
-        subscription_item_id = options.delete(:subscription_item_id) || subscription_items&.first&.dig("id")
+        subscription_item_id = options.delete(:subscription_item_id) || subscription_items&.first&.id
         if subscription_item_id
           ::Stripe::SubscriptionItem.update(subscription_item_id, options.merge(quantity: quantity), stripe_options)
           @api_record = nil
@@ -234,7 +237,7 @@ module Pay
         update(
           pause_behavior: behavior,
           pause_resumes_at: (@api_record.pause_collection&.resumes_at ? Time.at(@api_record.pause_collection&.resumes_at) : nil),
-          pause_starts_at: ((behavior == "void") ? Time.at(@api_record.current_period_end) : nil)
+          pause_starts_at: ((behavior == "void") ? Time.at(@api_record.items.first.current_period_end) : nil)
         )
       end
 
@@ -292,8 +295,8 @@ module Pay
         )
 
         # Validate that swap was successful and handle SCA if needed
-        if (payment_intent = @api_record.latest_invoice.payment_intent)
-          Pay::Payment.new(payment_intent).validate
+        if (payment_intent_id = @api_record.latest_invoice.payments.first.payment.payment_intent)
+          Pay::Payment.from_id(payment_intent_id).validate
         end
 
         sync!(object: @api_record)
@@ -301,33 +304,19 @@ module Pay
         raise Pay::Stripe::Error, e
       end
 
-      # Creates a metered billing usage record
-      #
-      # Uses the first subscription_item ID unless `subscription_item_id: "si_1234"` is passed
-      #
-      # create_usage_record(quantity: 4, action: :increment)
-      # create_usage_record(subscription_item_id: "si_1234", quantity: 100, action: :set)
-      def create_usage_record(**options)
-        subscription_item_id = options.delete(:subscription_item_id) || metered_subscription_item&.dig("id")
-        ::Stripe::SubscriptionItem.create_usage_record(subscription_item_id, options, stripe_options)
-      end
-
-      # Returns usage record summaries for a subscription item
-      def usage_record_summaries(**options)
-        subscription_item_id = options.delete(:subscription_item_id) || metered_subscription_item&.dig("id")
-        ::Stripe::SubscriptionItem.list_usage_record_summaries(subscription_item_id, options, stripe_options)
+      def subscription_items
+        stripe_object.items
       end
 
       # Returns the first metered subscription item
       def metered_subscription_item
-        subscription_items.find do |subscription_item|
-          subscription_item.dig("price", "recurring", "usage_type") == "metered"
+        subscription_items.auto_paging_each do |subscription_item|
+          return subscription_item if subscription_item.price.try(:recurring).try(:usage_type) == "metered"
         end
       end
 
-      # Returns an upcoming invoice for a subscription
-      def upcoming_invoice(**options)
-        ::Stripe::Invoice.upcoming(options.merge(subscription: processor_id), stripe_options)
+      def preview_invoice(**options)
+        ::Stripe::Invoice.create_preview(options.merge(subscription: processor_id), stripe_options)
       end
 
       # Retries the latest invoice for a Past Due subscription and attempts to pay it