+++ app/models/pay/subscription.rb
@@ -29,7 +29,6 @@ module Pay
 
     store_accessor :data, :paddle_update_url
     store_accessor :data, :paddle_cancel_url
-    store_accessor :data, :subscription_items
 
     attribute :prorate, :boolean, default: true