+++ db/migrate/20250415151129_add_object_to_pay_models.rb
@@ -1,0 +1,7 @@
+class AddObjectToPayModels < ActiveRecord::Migration[8.0]
+  def change
+    add_column :pay_charges, :object, Pay::Adapter.json_column_type
+    add_column :pay_customers, :object, Pay::Adapter.json_column_type
+    add_column :pay_subscriptions, :object, Pay::Adapter.json_column_type
+  end
+end