+++ lib/pay/stripe.rb
@@ -1,7 +1,6 @@
 module Pay
   module Stripe
     class Error < Pay::Error
-      delegate :message, to: :cause
     end
 
     module Webhooks
@@ -27,7 +26,7 @@ module Pay
 
     extend Env
 
-    REQUIRED_VERSION = "~> 13"
+    REQUIRED_VERSION = "~> 15"
 
     # A list of database model names that include Pay
     # Used for safely looking up models with client_reference_id