+++ lib/pay/stripe/webhooks/customer_updated.rb
@@ -11,6 +11,18 @@ module Pay
 
           stripe_customer = pay_customer.api_record
 
+          attributes = {
+            object: stripe_customer.to_hash
+          }
+
+          # Sync invoice credit balance and currency
+          if stripe_customer.invoice_credit_balance.present?
+            attributes[:invoice_credit_balance] = stripe_customer.invoice_credit_balance
+            attributes[:currency] = stripe_customer.currency
+          end
+
+          pay_customer.update(attributes)
+
           # Sync default card
           if (payment_method_id = stripe_customer.invoice_settings.default_payment_method)
             Pay::Stripe::PaymentMethod.sync(payment_method_id, stripe_account: event.try(:account))
@@ -18,14 +30,6 @@ module Pay
           else
             # No default payment method set
             pay_customer.payment_methods.update_all(default: false)
-          end
-
-          # Sync invoice credit balance and currency
-          if stripe_customer.invoice_credit_balance.present?
-            pay_customer.update(
-              invoice_credit_balance: stripe_customer.invoice_credit_balance,
-              currency: stripe_customer.currency
-            )
           end
         end
       end