+++ test/pay/stripe/webhooks/charge_refunded_test.rb
@@ -9,6 +9,7 @@ class Pay::Stripe::Webhooks::ChargeRefundedTest < ActiveSupport::TestCase
 
   test "a charge is updated with refunded amount" do
     ::Stripe::Charge.expects(:retrieve).returns(@event.data.object)
+    ::Stripe::InvoicePayment.expects(:list).returns([])
 
     assert_enqueued_jobs 1 do
       Pay::Stripe::Webhooks::ChargeRefunded.new.call(@event)