+++ test/pay/stripe/webhooks/charge_succeeded_test.rb
@@ -9,6 +9,7 @@ class Pay::Stripe::Webhooks::ChargeSucceededTest < ActiveSupport::TestCase
     pay_customers(:stripe).update(processor_id: @event.data.object.customer)
 
     ::Stripe::Charge.expects(:retrieve).returns(@event.data.object)
+    ::Stripe::InvoicePayment.expects(:list).returns([])
 
     # Make sure enqueues the receipt email
     assert_enqueued_jobs 1 do